searchable bar problem occured
