+++ Shared/MovieListPage/Tabs/AllMoviesView.swift
@@ -9,24 +9,27 @@ import SwiftUI
 
 struct AllMoviesView: View {
     // 9 - Buradan gitmeli
-    @ObservedObject var moviesViewModel: MainBackgroundViewModel
-    
+    //@ObservedObject var moviesViewModel: MainBackgroundViewModel
+    @ObservedObject var allMoviewViewModel = AllMoviesViewModel()
+
     var body: some View {
         ScrollView(showsIndicators: false){
             LazyVStack{
-                ForEach(moviesViewModel.searchResults.isEmpty ?  moviesViewModel.allMovies: moviesViewModel.searchResults, id: \.id) { movie in
+                //ForEach(moviesViewModel.searchResults.isEmpty ?  moviesViewModel.allMovies: moviesViewModel.searchResults, id: \.id) { movie in
+                ForEach(allMoviewViewModel.allMovies , id: \.id) { movie in
                     VStack{
                         Divider()
                         NavigationLink(destination: MovieDetailsView(viewModel: MovieDetailsViewModel(movie: movie))) {
                             MovieCellView(image: movie.image, title: movie.title, releaseDate: movie.release_date)
                         }
                         //Pagination
-                        if(moviesViewModel.searchResults.isEmpty && self.moviesViewModel.allMovies.last?.id == movie.id){
+                        //if(moviesViewModel.searchResults.isEmpty && self.moviesViewModel.allMovies.last?.id == movie.id){
+                        if(self.allMoviewViewModel.allMovies.last?.id == movie.id){
                             Divider()
                             Text("Fetching more...")
                                 .onAppear(perform: {
                                     // 10 - Burası metod olmalı sadece
-                                    self.moviesViewModel.getTopMovies(pageNum: self.moviesViewModel.updateMovies())
+                                    self.allMoviewViewModel.getTopMovies(pageNum: self.allMoviewViewModel.updateMovies())
                                 })
                         }
                     }
@@ -38,6 +41,6 @@ struct AllMoviesView: View {
 
 struct AllMoview_Previews: PreviewProvider {
     static var previews: some View {
-        AllMoviesView(moviesViewModel: MainBackgroundViewModel())
+        AllMoviesView()
     }
 }
